+++ src/collector/mod.rs
@@ -175,12 +175,14 @@ pub trait Collector: Sync + Send {
         if let Some(alive_bitset) = reader.alive_bitset() {
             weight.for_each(reader, &mut |doc, score| {
                 if alive_bitset.is_alive(doc) {
-                    segment_collector.collect(doc, score).unwrap(); // TODO
+                    segment_collector.collect(doc, score)?;
                 }
+                Ok(())
             })?;
         } else {
             weight.for_each(reader, &mut |doc, score| {
-                segment_collector.collect(doc, score).unwrap(); // TODO
+                segment_collector.collect(doc, score)?;
+                Ok(())
             })?;
         }
         Ok(segment_collector.harvest())