+++ src/query/weight.rs
@@ -7,13 +7,14 @@ use crate::{DocId, Score, TERMINATED};
 /// `DocSet` and push the scored documents to the collector.
 pub(crate) fn for_each_scorer<TScorer: Scorer + ?Sized>(
     scorer: &mut TScorer,
-    callback: &mut dyn FnMut(DocId, Score),
-) {
+    callback: &mut dyn FnMut(DocId, Score) -> crate::Result<()>,
+) -> crate::Result<()> {
     let mut doc = scorer.doc();
     while doc != TERMINATED {
-        callback(doc, scorer.score());
+        callback(doc, scorer.score())?;
         doc = scorer.advance();
     }
+    Ok(())
 }
 
 /// Calls `callback` with all of the `(doc, score)` for which score
@@ -71,10 +72,10 @@ pub trait Weight: Send + Sync + 'static {
     fn for_each(
         &self,
         reader: &SegmentReader,
-        callback: &mut dyn FnMut(DocId, Score),
+        callback: &mut dyn FnMut(DocId, Score) -> crate::Result<()>,
     ) -> crate::Result<()> {
         let mut scorer = self.scorer(reader, 1.0)?;
-        for_each_scorer(scorer.as_mut(), callback);
+        for_each_scorer(scorer.as_mut(), callback)?;
         Ok(())
     }